data-processing/requirements.txt
### Dependencies ###
google-cloud-datastore==1.3.0